requirement.txt
asgiref==3.7.2
captcha==0.5.0
contourpy==1.1.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-notification==0.2
django-crispy-forms==2.0
django-forms-bootstrap==3.1.0
django-import-export==3.2.0
et-xmlfile==1.1.0
fabio==2023.6.0
fonttools==4.42.0
h5py==3.9.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.7.2
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
PyMySQL==1.1.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
PyYAML==6.0.1
silx==1.1.2
six==1.16.0
sqlparse==0.4.4
style==1.1.0
tablib==3.5.0
update==0.0.1
xlrd==2.0.1
xlwt==1.3.0
